[yamllint] Make yamllint pass with indent-sequence: consistent
Picking either "true" (always indent sequence elements) or "false" (never indent
sequence elements) results in massive failures, with some apparently-generated
files disagreeing in their conventions. But making the world pass with "consistent"
(indent or don't, but be consistent within a document) requires only changing
things that are written by hand... a lot of things, to be sure, but we should
need this kind of edit only once.

Signed-off-by: Flynn <emissary@flynn.kodachi.com>