Fix `scopelint` complaints [ci-skip]
`scopelint` is a little over-zealous; If a sub-test doesn't call
`t.Parallel()`, it isn't *really* necessary to address scopelint's
complaint, but it doesn't hurt.

This reveals a few test failures that weren't being caught because the
testcase was getting reset as it was running; those will be fixed in
the next commit.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>