Turn on the `scopelint` linter, fix complaints (#365)
* Turn on the `scopelint` linter [ci-skip]

While reviewing https://github.com/envoyproxy/gateway/pull/356 I noticed a
common mistake: https://github.com/golang/go/wiki/CommonMistakes#using-goroutines-on-loop-iterator-variables

Then a quick `grep` turned up several existing places that the mistake has
already been made.  So I thought to myself "I bet there's a linter we can
turn on to catch this."

As far as I can tell, the best match is `scopelint`, which is a little
over-zealous.  If a sub-test doesn't call `t.Parallel()`, it isn't
*really* necessary to address scopelint's complaint, but it doesn't hurt.

This reveals a bunch of places with bad scopes that `scopelint` will
complain about; fix those in the next commit.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>

* Fix `scopelint` complaints [ci-skip]

`scopelint` is a little over-zealous; If a sub-test doesn't call
`t.Parallel()`, it isn't *really* necessary to address scopelint's
complaint, but it doesn't hurt.

This reveals a few test failures that weren't being caught because the
testcase was getting reset as it was running; those will be fixed in
the next commit.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>

* Fix newly discovered test failures

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>