set Gateway namespace and name labels on dependent resources
Switches from a single label that only specifies
the Gateway name to two labels, one for namespace
and one for name, on Gateway dependent resources.
This avoids conflicts between Gateways in different
namespaces with the same names.

Also fixes mapping from dependent resources to
the owning Gateway in the Gateway reconciler.

Closes #452.
Closes #453.

Signed-off-by: Steve Kriss <krisss@vmware.com>