Fix Service object remove from resource map from route controller
This commit fetches the services that were referenced by the route
which has now been deleted, and removes those objects from the
resource map. In order to facilitate this, the commit introduces
a provider cache for kubernetes that stores mappings between
kubernetes objects - for now it stores the map between
[tls/http]route -> backend service references.
So that once the [tls/http]route is deleted we have the services that
it referenced.

Signed-off-by: Shubham Chauhan <shubham@tetrate.io>