Fix intermittent failures (#557)
* .gitignore: Ignore `vendor/` directories

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>

* xds translator: Fix racy startup

If the watchable.Map has content in it already when .Subscribe() is called
on it, then those initial entries won't have a snapshot.Updates entry in
that first snapshot.  For the first snapshot we just need to iterate over
snapshot.State.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>

* provider tests: Fix running the test multiple times

controller-runtime.SetupSignalHandler() panics if called more than once in
a process.  So running the test multiple times (`go test -count=2`)
reliably causes the test to panic.

So don't use ctrl.SetupSignalHandler() in unit tests.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>

* Add and use a new watchutil.HandleSubscription function

As the added godoc comment says, "This is better than iterating over
snapshot.Updates because it handles the case where the the watchable.Map
already contains entries before .Subscribe is called."

The generalizes the fix that I made in the XDS translator.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>

* docs: Add a bit to watching.md about HandleSubscription

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>

* Move HandleSubscription et al. around per Arko's feedback

I was going to do a type alias for `watchable.Update`, but:

    internal/message/watchutil.go:7:6: generic type cannot be alias

So I just defined a new child type, which is fine because there aren't any
methods on Update.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>