translator: remove usage of deprecated Http2ProtocolOptions
Envoy deprecated the usage of `Http2ProtocolOptions` in Envoy 1.17.
This commit updates the xds.Translate function to follow recommended
practices of using the `TypedExtensionProtocolOptions` to set
http/2 protocol options.

The existing test case was modified due to the change in expected xDS
configuration outputted when http/2 is enabled for a backendRef.

Signed-off-by: Lance Austin <laustin@datawire.io>