gateway-api: fix supported route kind validation
The logic for validating `allowedRoute.kinds` of a listener was
overwriting the SupportedKinds when multiple kinds are configured.
This caused the `GatewayInvalidRouteKind` conformance test to fail.

Items addressed:
- refactored `validateAllowedRoutes` to reduce nesting
- fixed `validateAllowedRoutes` logic so  it collects all of the
SupportedKinds before setting them on the listener
- provide additional context to user to help identify which RouteKind
is invalid on a listener
- updated unit tests for translator based on new expected output
- added additional unit test case for translator to test for when
multiple Kinds are added which includes valid and invalid Kinds.

Signed-off-by: Lance Austin <laustin@datawire.io>