Changes HTTPRoute hostname matching to ignore trailing port numbers.
This is useful for testing behind NAT, e.g. when deploying in a VM.

HTTP authority is e.g. 'example.com:1234', but the Gateway API
restricts the values of Hostnames to domain names. There are two
interpretations: (1) port numbers are not allowed or (2) only the host
portion should be matched. (1) is not a problem if the cluster owner
only uses the standard ports, so (2) seems like a more useful
interpretation.

This cannot use a simple `contains_match`, since the separating colon
is optional. For performance, perhaps exact matches should be
special-cased somehow: that ignoring the port is only done for routes
with a certain annotation.