Fix docs for `ExtensionTLS` type
Current document says to mount a secret that has a TLS private key. This is incorrect - Envoy Gateway acting as a client should not receive any private key.

Envoy Gateway doesn't support mTLS when connecting to extension server, so there is no need for private key today. This is verified by reading the code. EG is only looking for `tls.crt`
 https://github.com/envoyproxy/gateway/blob/28e1a485fad395291dd633e3d472f130455de0e8/internal/extension/registry/extension_manager.go#L265-L276

Cert-manager docs explain this concept concisely, so link directly to it.

Signed-off-by: Teju Nareddy <tejunareddy@gmail.com>