Adding SAP as an adopter. Notice:
- Use of the SAP logo is limited solely to the Purpose of showing SAP as an adopter/user of this open source project
- Use of the SAP logo must conform to [SAP’s trademark use guidelines](https://www.sap.com/corporate/en/legal/trademark.html#proper-use)
- The context in which the SAP logo is displayed shall be positive to SAP and its products and services

SAP hereby consents to the use of its logo for the limited Purpose stated above and Company’s use of the logo will serve as its acknowledgment of the accuracy of the representations upon which this consent relies. SAP may revoke this consent at any time and without notice if Company violates any of the representations or conditions contained in this message. SAP may revoke this notice under any other circumstances with thirty (30) days notice to you.

SAP reserves all intellectual property rights in and to its trademarks. Company’s use of the SAP trademarks in violation of this consent letter may be subject to legal action under applicable law. SAP’s consent and the logos and/or trademarks to which the consent applies are provided to Company as-is with no warranties of any type

Signed-off-by: Guy Daich <guy.daich@sap.com>