fix(translator): make OIDC and JWT authentication work together (#5142)
* Make OIDC and JWT authentication work together

Currently, when both OIDC and JWT authentication mechanisms are configured in
the same SecurityPolicy, the OIDC is applied first. It ensures the presence of
the tokens in the request's cookies, and adds the Authorisation header to the
request. Then JWT is applied, validating the added header.

This setup works perfectly for browser requests. However, it blocks requests
from clients that provide a valid "Authorization: Bearer..." header (normally
non-browser clients). The OIDC mechanism kicks in first and redirects the
requests to the login pages because of the missing cookies.

Envoy Proxy has a pass_through_matcher option that can be used to skip over the
OIDC filter when the request is going to be handled by the JWT filter later.
Add a new option passThroughAuthHeader on the OIDC config that will add this
option in xDS.

It would be an error to have the new passThroughAuthHeader enabled without also
having JWT config that will later check the validity of the token. If this is
not the case, the SecurityPolicy is rejected for safety.

Note that this change is designed to be entirely backwards compatible. The
system behaviour is unchanged unless the new option is switched on.

Signed-off-by: Stephen Robin <stephen.robin@gmail.com>

* Add an e2e test of OIDC and JWT authentication used together

Signed-off-by: Stephen Robin <stephen.robin@gmail.com>