standalone: use the right version of Envoy (#6891)
When using the standalone mode and the Host infrastructure, EG uses func-e to download and/or start Envoy, but the version of Envoy to run is not explicitly set. This could lead to undersider behaviour if content already exists in the /tmp/envoy-gateway directory and EG may not use the expected version.

This PR adds a change to:

Infer the version to use from the default Envoy Image tag. This is only set on the release branches, but not main. In that case, fallback to the old behaviour for backwards compatibility and let func-e decide the version to use.
Ad a config knob to the EnvoyProxy API to allow setting the version of Envoy to use when in standalone mode:
spec:
  provider:
    type: Host
    host:
      envoyVersion: "1.35.0"
---------

Signed-off-by: Ignasi Barrera <nacx@apache.org>