fix(translator): matching upstream http(s) scheme
This makes it possible to use gateways that are internally exposed to
Kubernetes as an HTTP service that proxies to a strictly HTTPS service.
Some HTTPS services will declare a protocol error when an HTTPS request
is forwarded towards them but the HTTP2 `:scheme` header is still set to
`http`.

Signed-off-by: Andreas Wachs <awa@corti.ai>